Isolate service logic from the views for authentication
